Move the tests to de dedicated folders to prepare for the big split #10
