+++ phpunit.xml.dist
@@ -1,15 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<phpunit bootstrap="./tests/bootstrap.php" colors="true">
+<phpunit bootstrap="./vendor/autoload.php" colors="true">
     <testsuites>
         <testsuite>
-            <directory suffix="Test.php">./tests</directory>
+            <directory suffix="Test.php">./src/Bundle/LongRunningBundle/Tests/</directory>
+            <directory suffix="Test.php">./src/Core/Tests/</directory>
+            <directory suffix="Test.php">./src/Plugin/*/Tests/</directory>
         </testsuite>
     </testsuites>
     <filter>
         <whitelist>
             <directory>./</directory>
             <exclude>
-                <directory>./tests</directory>
                 <directory>./vendor</directory>
             </exclude>
         </whitelist>
