+++ src/Bundle/LongRunningBundle/Tests/Functional/MonologBufferTest.php
@@ -1,18 +1,17 @@
 <?php
 
-namespace LongRunning\Tests\Functional;
+namespace LongRunning\Bundle\LongRunningBundle\Tests\Functional;
 
 use LongRunning\Core\DelegatingCleaner;
-use Monolog\Handler\FingersCrossedHandler;
 use Monolog\Handler\TestHandler;
 use Psr\Log\LoggerInterface;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 
-class MonologFingersCrossedTest extends KernelTestCase
+class MonologBufferTest extends KernelTestCase
 {
     protected static function getKernelClass()
     {
-        return 'LongRunning\Tests\Functional\TestKernel';
+        return 'LongRunning\Bundle\LongRunningBundle\Tests\Functional\TestKernel';
     }
 
     protected function setUp()
@@ -23,25 +22,21 @@ class MonologFingersCrossedTest extends KernelTestCase
     /**
      * @test
      */
-    public function it_clears_messages_that_never_hit_the_action_level()
+    public function it_keeps_messages_in_buffer_until_cleanp()
     {
-        $bufferTestHandler = $this->getTestHandler('fingers_crossed_test');
+        $bufferTestHandler = $this->getTestHandler('buffer_test');
         $logger = $this->getLogger();
 
-        $logger->debug('This message will never show up');
+        $logger->debug('Buffer this message');
         $this->assertMessages($bufferTestHandler->getRecords(), array());
 
         $this->getCleaner()->cleanUp();
 
-        $this->assertMessages($bufferTestHandler->getRecords(), array());
-
-        $fingersCrossedHandler = $this->getFingersCrossedHandler('fingers_crossed');
-        $reflectionObject = new \ReflectionObject($fingersCrossedHandler);
-        $property = $reflectionObject->getProperty('buffer');
-        $property->setAccessible(true);
-
-        $messages = $property->getValue($fingersCrossedHandler);
-        $this->assertMessages($messages, array(
+        $this->assertMessages($bufferTestHandler->getRecords(), array(
+            'Buffer this message',
+            'Clear EntityManager',
+            'Close database connection',
+            'Clear monolog fingers crossed handler',
             'Close monolog buffer handler',
         ));
     }
@@ -67,15 +62,6 @@ class MonologFingersCrossedTest extends KernelTestCase
      * @return TestHandler
      */
     private function getTestHandler($name)
-    {
-        return static::$kernel->getContainer()->get(sprintf('monolog.handler.%s', $name));
-    }
-
-    /**
-     * @param string $name
-     * @return FingersCrossedHandler
-     */
-    private function getFingersCrossedHandler($name)
     {
         return static::$kernel->getContainer()->get(sprintf('monolog.handler.%s', $name));
     }
